refactor: selector name change
