+++ src/components/DataDimension/styles/DetailSelector.style.js
